+++ CHANGELOG.md
@@ -23,6 +23,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Update the label of "Copy as PowerShell Invoke-WebRequest command" to confine supporting PowerShell version 6 or above only (there is no change to the underlying logic)
 - Inherited values in Request Editor are now showing at the bottom rather than the top
 - Number of space characters to indent or unindent in Request Body and Payload editors are changed from 4 to 2
+- Syntax highlighting would be disabled for Request/Payload text fields that exceeding 1.5 MB size
+- Syntax highlighting would be disabled for Response body that exceeding 10 MB size
 
 ### Removed
 - Text fields and response body viewer now do not trim content over 4 MB (but other limits still apply)